requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator>=2.1.0
python-dotenv==1.0.0
supabase==2.25.1
werkzeug==3.0.1

